Fixed issues found by Coverity
The resource leak was real, if unlikely. I don't believe the NULL issues
in from_map and from_array actually exist. I think the NULL checker
thinks there may be an issue because of the NULL check in the for loop.
However, I moved a null check from the Reader_get function to the
from_entry_data_list function. If entry_data_list is NULL or the pointer
that it points to is NULL, it will now be caught there before the
from_map and from_array calls.