Improve type checking: return ‘Reader’ from open_database()
This exposes the ‘Reader’ class under its own name (instead of an alias)
and uses it as the return type for open_database(). In case the C
extension is used, cast that to the same ‘Reader’ type for type checking
purposes, since they share an identical API.

The benefits are that this avoids an unnecessary Union return type
(see e.g. https://github.com/python/mypy/issues/1693) and doesn't
require internal types to be exposed with confusing names for type
checking purposes only.

The ‘Reader’ compat helper is unnecessary and can be dropped: the Python
Reader is always available and API compatible.

While at it, simplify and isort the imports.