Run AddressSanitizer
I wanted to use Clang for this but I've been having issues getting the
final step of the compilation to use Clang. It uses Clang for building
and GCC for linking. That might be fine and the correct thing to do in
fact, but I'm not sure. Either way, using ASAN with GCC seems
sufficient.