Preload libasan
For some reason we had begun not being able to load the extension. The
message when loading the extension was:

ASan runtime does not come first in initial library list; you should
either link runtime to your application or manually preload it with
LD_PRELOAD.

This loads libasan now. We have to disable leak checking at the same
time as otherwise we see leak errors from outside our code.