Fix multiprocessing tests on macOS.
On macOS, the default method for creating a new process is to use spawn
instead of fork. When this occurs, arguments to the process target need
to be pickled, but the test includes several unpicklable objects, such
as local functions and mmaps.

We can fix the test's current assumptions by forcing "fork" instead of
"spawn".