Correctly handle null check of entry_data_list
I suspect at some point, we were passing in `entry_data_list` as a
`MMDB_entry_data_list_s *`, which is what we do in other extensions, but
that was refactored to `MMDB_entry_data_list_s **` for reasons I do not
recall, and the loop check was never updated appropriately.

This never caused an actual issue as we also have the size check.
However, it could potentially lead to a segfault on a corrupt or
otherwise invalid database.